+++ packages/core/src/models/constants.ts
@@ -5,7 +5,7 @@ export const constants = {
   lockFileDestination: ".Leapp/Leapp-lock.json",
   lockFileBackupPath: ".Leapp/Leapp-lock.backup.bin",
   latestUrl: "https://leapp.cloud/releases.html",
-  workspaceLastVersion: 5,
+  workspaceLastVersion: 6,
   slackUrl: "https://join.slack.com/t/noovolari/shared_invite/zt-opn8q98k-HDZfpJ2_2U3RdTnN~u_B~Q",
   localWorkspaceName: "Local workspace",
   localWorkspaceDescription: "Community Edition",
@@ -83,4 +83,11 @@ export const constants = {
     "yP6Ekgn8yeDmON1JoX0E0Tdm6dfXVY5v2K8KaT2/XtsqRlwCCU+pXRsQwJfQcknZ\n" +
     "aNqnY3aBtCwmaesTlrc5bR8CAwEAAQ==\n" +
     "-----END PUBLIC KEY-----",
+
+  touchIdKeychainItemName: "touch-id-lock",
+  touchIdEnabled: true,
+  requirePasswordEveryWeek: { key: "Every week", value: 7 },
+  requirePasswordEveryTwoWeeks: { key: "Every 2 weeks", value: 14 },
+  requirePasswordEveryMonth: { key: "Every month", value: 30 },
+  requirePasswordNever: { key: "Never", value: 999999 },
 };