+++ packages/core/src/models/workspace.ts
@@ -37,6 +37,9 @@ export class Workspace {
   private _colorTheme: string;
   private _extensionEnabled: boolean;
 
+  private _requirePassword: number;
+  private _touchIdEnabled: boolean;
+
   private _proxyConfiguration: {
     proxyProtocol: string;
     proxyUrl?: string;
@@ -242,5 +245,21 @@ export class Workspace {
 
   set notifications(notifications: LeappNotification[]) {
     this._notifications = notifications;
+  }
+
+  get requirePassword(): number {
+    return this._requirePassword;
+  }
+
+  set requirePassword(value: number) {
+    this._requirePassword = value;
+  }
+
+  get touchIdEnabled(): boolean {
+    return this._touchIdEnabled;
+  }
+
+  set touchIdEnabled(value: boolean) {
+    this._touchIdEnabled = value;
   }
 }