+++ packages/desktop-app/src/app/components/dialogs/authorization-dialog/authorization-dialog.component.ts
@@ -1,0 +1,20 @@
+import { Component, Input, OnInit } from "@angular/core";
+import { BsModalRef } from "ngx-bootstrap/modal";
+
+@Component({
+  selector: "app-authorization-dialog",
+  templateUrl: "./authorization-dialog.component.html",
+  styleUrls: ["./authorization-dialog.component.scss"],
+})
+export class AuthorizationDialogComponent implements OnInit {
+  @Input()
+  authorizationCode: string;
+
+  constructor(private bsModalRef: BsModalRef) {}
+
+  ngOnInit(): void {}
+
+  close(): void {
+    this.bsModalRef.hide();
+  }
+}