+++ packages/desktop-app/src/app/components/dialogs/leapp-pro-pre-checkout-dialog/leapp-pro-pre-checkout-dialog.component.ts
@@ -24,14 +24,17 @@ export class LeappProPreCheckoutDialogComponent implements OnInit {
 
   public emailFormControl = new FormControl("", [Validators.required, Validators.email]);
   public planFormControl = new FormControl("annually");
+  // public fiscalCodeFormControl = new FormControl("", [Validators.pattern(/^[A-Za-z]{6}[0-9]{2}[A-Za-z]{1}[0-9]{2}[A-Za-z]{1}[0-9]{3}[A-Za-z]{1}$/)]);
 
   public form = new FormGroup({
     email: this.emailFormControl,
     plan: this.planFormControl,
+    // fiscalCode: this.fiscalCodeFormControl,
   });
 
   public selectedPeriod: BillingPeriod = BillingPeriod.yearly;
   public isEmailValid = false;
+  // public isCFValid = false;
   public price: any;
   private prices: any[];
 
@@ -65,11 +68,15 @@ export class LeappProPreCheckoutDialogComponent implements OnInit {
   }
 
   async upgradeToLeappPro(): Promise<void> {
-    if (this.isEmailValid) {
+    if (this.isEmailValid /* && this.isCFValid*/) {
       let checkoutUrl = "";
       try {
         this.isRedirectingToCheckout = true;
-        checkoutUrl = await this.appProviderService.teamService.createCheckoutSession(this.emailFormControl.value, this.price);
+        checkoutUrl = await this.appProviderService.teamService.createCheckoutSession(
+          this.emailFormControl.value,
+          this.price
+          // this.fiscalCodeFormControl?.value
+        );
         this.isRedirectingToCheckout = false;
       } catch (error) {
         if (error.response.data?.errorCode === ApiErrorCodes.emailAlreadyTaken) {
@@ -133,6 +140,9 @@ export class LeappProPreCheckoutDialogComponent implements OnInit {
 
   checkAndConfirm(): void {
     this.emailFormControl.markAsTouched();
-    this.isEmailValid = this.form.valid;
+    this.isEmailValid = this.emailFormControl.valid;
+
+    // this.fiscalCodeFormControl.markAsTouched();
+    // this.isCFValid = this.fiscalCodeFormControl.valid;
   }
 }