+++ packages/desktop-app/src/app/components/dialogs/login-team-dialog/login-workspace-dialog.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from "@angular/core";
 import { AppService } from "../../../services/app.service";
+import { OptionsService } from "../../../services/options.service";
 import { AppProviderService } from "../../../services/app-provider.service";
 import { LoggedEntry, LogLevel, LogService } from "@noovolari/leapp-core/services/log-service";
-import { TeamService, ApiErrorCodes, FormErrorCodes } from "../../../services/team-service";
+import { ApiErrorCodes, FormErrorCodes, TeamService } from "../../../services/team-service";
 
 import { AbstractControl, FormControl, FormGroup, Validators } from "@angular/forms";
 import { globalLeappProPlanStatus, LeappPlanStatus } from "../options-dialog/options-dialog.component";
@@ -22,7 +23,7 @@ export class LoginWorkspaceDialogComponent implements OnInit {
   private loggingService: LogService;
   private teamService: TeamService;
 
-  constructor(public appService: AppService, public appProviderService: AppProviderService) {
+  constructor(public appService: AppService, public appProviderService: AppProviderService, public optionsService: OptionsService) {
     this.email = new FormControl("", [Validators.required, Validators.email]);
     this.password = new FormControl("", [Validators.required]);
     this.signinForm = new FormGroup({ email: this.email, password: this.password });
@@ -45,9 +46,13 @@ export class LoginWorkspaceDialogComponent implements OnInit {
         const signedInUser = await this.teamService.signedInUserState.getValue();
         const doesWorkspaceExist = !!signedInUser;
         await this.teamService.signIn(formValue.email, formValue.password);
+        await this.teamService.writeTouchIdCredentials(formValue.password, this.optionsService.requirePassword);
         this.appService.closeAllMenuTriggers();
-        await this.appProviderService.keychainService.saveSecret("Leapp", "leapp-enabled-plan", LeappPlanStatus.proEnabled);
-        globalLeappProPlanStatus.next(LeappPlanStatus.proEnabled);
+
+        const teamOrPro = this.teamService.workspacesState.getValue().find((wState) => wState.type === "pro" || wState.type === "team");
+        const planStatus = teamOrPro.type === "team" ? LeappPlanStatus.enterprise : LeappPlanStatus.proEnabled;
+        await this.appProviderService.keychainService.saveSecret("Leapp", "leapp-enabled-plan", planStatus);
+        globalLeappProPlanStatus.next(planStatus);
         this.closeModal();
         if (doesWorkspaceExist) {
           await this.teamService.pullFromRemote();
@@ -55,12 +60,13 @@ export class LoginWorkspaceDialogComponent implements OnInit {
           this.loggingService.log(new LoggedEntry(`Welcome ${formValue.email}!`, this, LogLevel.success, true));
         }
       } catch (responseException: any) {
-        if (responseException?.response.data?.errorCode === ApiErrorCodes.invalidCredentials) {
+        if (responseException?.response?.data?.errorCode === ApiErrorCodes.invalidCredentials) {
           this.loggingService.log(new LoggedEntry("Invalid email or password", this, LogLevel.error, true));
           this.password.setErrors({ [FormErrorCodes.invalidCredentials]: {} });
-        } else if (responseException?.response.data?.errorCode === ApiErrorCodes.userNotActive) {
+        } else if (responseException?.response?.data?.errorCode === ApiErrorCodes.userNotActive) {
           this.loggingService.log(new LoggedEntry("The user is not active", this, LogLevel.error, true));
         } else {
+          console.log("ECCEZIONE", responseException);
           this.loggingService.log(new LoggedEntry(responseException, this, LogLevel.error, true));
         }
       } finally {