+++ packages/desktop-app/src/app/components/sync-pro-widget/sync-pro-widget.component.spec.ts
@@ -1,6 +1,9 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { SyncProWidgetComponent } from "./sync-pro-widget.component";
+import { MAT_SNACK_BAR_DATA, MatSnackBarRef } from "@angular/material/snack-bar";
+import { mustInjected } from "../../../base-injectables";
+import { RouterTestingModule } from "@angular/router/testing";
 
 describe("SyncProWidgetComponent", () => {
   let component: SyncProWidgetComponent;
@@ -9,6 +12,11 @@ describe("SyncProWidgetComponent", () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [SyncProWidgetComponent],
+      imports: [RouterTestingModule],
+      providers: [
+        { provide: MAT_SNACK_BAR_DATA, useValue: {} },
+        { provide: MatSnackBarRef, useValue: {} },
+      ].concat(mustInjected()),
     }).compileComponents();
   });