+++ packages/desktop-app/src/app/services/options.service.ts
@@ -148,4 +148,26 @@ export class OptionsService {
     workspace.samlRoleSessionDuration = value * 60 * 60;
     this.workspaceService.persistWorkspace(workspace);
   }
+
+  get requirePassword(): number {
+    const workspace = this.workspaceService.getWorkspace();
+    return workspace.requirePassword;
+  }
+
+  set requirePassword(value: number) {
+    const workspace = this.workspaceService.getWorkspace();
+    workspace.requirePassword = value;
+    this.workspaceService.persistWorkspace(workspace);
+  }
+
+  get touchIdEnabled(): boolean {
+    const workspace = this.workspaceService.getWorkspace();
+    return workspace.touchIdEnabled;
+  }
+
+  set touchIdEnabled(value: boolean) {
+    const workspace = this.workspaceService.getWorkspace();
+    workspace.touchIdEnabled = value;
+    this.workspaceService.persistWorkspace(workspace);
+  }
 }