+++ packages/desktop-app/src/app/services/team-service-stub.ts
@@ -28,7 +28,7 @@ export interface WorkspaceState {
 }
 
 export class TeamService {
-  constructor(_l: any, _e: any, _a: any, _p: any, _t: any, _m: any, _s: any, _r: any, _v: any, _i: any, _c: any, _n: any, _o: any) {}
+  constructor(_l: any, _e: any, _a: any, _p: any, _t: any, _m: any, _s: any, _r: any, _v: any, _i: any, _c: any, _n: any, _o: any, _y: any) {}
 
   get isLeappTeamStubbed(): boolean {
     return true;
@@ -65,6 +65,8 @@ export class TeamService {
   async signIn(_: string, __: string): Promise<void> {}
 
   async signOut(_: boolean = false): Promise<void> {}
+
+  async writeTouchIdCredentials(_: string, __: number): Promise<void> {}
 
   async deleteTeamWorkspace(): Promise<void> {}